+++ src/game/tests.rs
@@ -532,6 +532,7 @@ mod banding_blocked;
 mod banding_formation;
 mod banding_with_other;
 mod blaze_of_glory;
+mod block_restriction_cards;
 mod blocking_after_death;
 mod blocking_prevention;
 mod blocking_relation;